git-svn-id: svn://132.239.158.177/projects/ratrix/trunk@1282 d291aab3-510b-4049-a6d1-fe65b8d52d70

+++ analysis/eflister/compileTrialRecords.m
@@ -190,7 +190,7 @@ for i=1:length(subjectFiles)
 
 
                 %factor out trial range check
-                theseTrials=[1:length(newTrialRecs)];
+                theseTrials=int32([1:length(newTrialRecs)]);
                 shouldBe=ranges{i}(1,j)+theseTrials-1 <= ranges{i}(2,j);
                 if ~all(shouldBe)
                     shouldBe
